process indirect internal crossreferences: https://github.com/metanorma/metanorma/issues/122
