+++ spec/fixtures/collection/dummy.xml
@@ -90,8 +90,10 @@ at the address below or ISO’s member body in the country of the requester.
   
 </boilerplate>
 <sections><clause id="_scope" obligation="normative"><title>Scope</title><p id="_9f7c3e34-6854-4fa6-bb3b-e967ff9f5bf2">This is a dummy document.</p>
+  </clause>
+  <clause id="E1" type="express-schema">
 <figure id="_b4013f8f-ee40-4186-a5ed-3a5103f199e1">
-<pre id="_0855e974-d21b-48ad-8627-0c53719b8dcf">
+<pre id="E1.A1">
 module Metanorma
   class Document
     # @return [Strin]
@@ -181,7 +183,9 @@ end
 <p id="_5bbe5050-8dc1-4b7b-936c-d63ca93fa72a">What?</p>
 </li>
 </ol>
-<figure id="_c70d742f-3b61-4aaa-8343-981678eeb1f4">
+</clause>
+  <clause id="E2" type="express-schema">
+<figure id="E2.A2">
 <pre id="_923365c0-4c6d-473e-bb0e-e2ee87dcc438">
 module Metanorma
   # Metanorma collection's manifest
